move load and save messages to ShipyardSystem.Triad.Consoles
also many new generic methods in the shipyard systems
